Revert "Chore: Flatten NotFoundPage and UseCasePage"
This reverts commit 5ab2f0a5a88acc452952c668243bffdb6ae86881.
